project/tools/requirements.txt
# Gunicorn==19.9.0

psycopg2
django==3.0.3
djangorestframework
django-environ
django-model-utils
Pillow
